requirements.txt
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
faiss-cpu>=1.7.4
numpy>=1.24.0
linkup-sdk>=0.1.0

